Voters and code upgrade

+++ src/Controller/TaskController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller;
 
+use DateTime;
 use App\Entity\Task;
 use App\Entity\User;
 use App\Form\TaskType;
@@ -16,19 +17,19 @@ use Symfony\Component\Security\Http\Attribute\IsGranted;
 class TaskController extends AbstractController
 {
 
-    #[Route("/tasksDone", name: "task_list_done")]
+    #[Route("/tasksDone", name: "homepage_done", methods: ['GET'])]
     public function listDone(TaskRepository $taskRepository)
     {
         return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findByIsDoneTrue($this->getUser())]);
     }
 
-    #[Route("/tasksNotDone", name: "task_list_notdone")]
+    #[Route("/tasksNotDone", name: "homepage_notdone", methods: ['GET'])]
     public function listNotDone(TaskRepository $taskRepository)
     {
         return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findByIsDoneFalse($this->getUser())]);
     }
 
-    #[Route("/tasks/create", name: "task_create")]
+    #[Route("/tasks/create", name: "task_create", methods: ['GET', 'POST'])]
     public function create(Request $request, EntityManagerInterface $em)
     {
         $task = new Task();
@@ -37,20 +38,24 @@ class TaskController extends AbstractController
         $form->handleRequest($request);
 
         if($form->isSubmitted() && $form->isValid()) {
+            $task->setUser($this->getUser());
+            $task->setCreatedAt(new \DateTime());
+            $task->setIsDone(false);
             $em->persist($task);
             $em->flush();
 
             $this->addFlash('success', 'La tâche a été bien été ajoutée.');
 
-            return $this->redirectToRoute('task_list');
+            return $this->redirectToRoute('homepage');
         }
 
         return $this->render('task/create.html.twig', ['form' => $form->createView()]);
     }
 
-    #[Route("/tasks/{id}/edit", name: "task_edit")]
+    #[Route("/tasks/{id}/edit", name: "task_edit", methods: ['GET', 'POST'])]
     public function edit(Task $task, Request $request, EntityManagerInterface $em)
     {
+        $this->denyAccessUnlessGranted('edit', $task,  $translator->trans("Access denied, you can't edit this task"));
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
@@ -60,7 +65,7 @@ class TaskController extends AbstractController
 
             $this->addFlash('success', 'La tâche a bien été modifiée.');
 
-            return $this->redirectToRoute('task_list');
+            return $this->redirectToRoute('homepage');
         }
 
         return $this->render('task/edit.html.twig', [
@@ -69,25 +74,27 @@ class TaskController extends AbstractController
         ]);
     }
 
-    #[Route("/tasks/{id}/toggle", name: "task_toggle")]
+    #[Route("/tasks/{id}/toggle", name: "task_toggle", methods: ['GET', 'POST'])]
     public function toggleTask(Task $task, EntityManagerInterface $em)
     {
+        $this->denyAccessUnlessGranted('edit', $task,  $translator->trans("Access denied, you can't toggle this task"));
         $task->toggle(!$task->isDone());
         $em->flush();
 
         $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
 
-        return $this->redirectToRoute('task_list');
+        return $this->redirectToRoute('homepage');
     }
 
-    #[Route("/tasks/{id}/delete", name: "task_delete")]
+    #[Route("/tasks/{id}/delete", name: "task_delete", methods: ['POST'])]
     public function deleteTask(Task $task, EntityManagerInterface $em)
     {
+        $this->denyAccessUnlessGranted('delete', $task,  $translator->trans("Access denied, you can't delete this task"));
         $em->remove($task);
         $em->flush();
 
         $this->addFlash('success', 'La tâche a bien été supprimée.');
 
-        return $this->redirectToRoute('task_list');
+        return $this->redirectToRoute('homepage');
     }
 }
